Refactoring integers

+++ dlls/ga.cpp
@@ -177,7 +177,7 @@ ga_value CPopulation::bestFitness() const
 
 ga_value CPopulation::averageFitness() const
 {
-	return totalFitness() / m_theIndividuals.size();
+	return totalFitness() / static_cast<float>(m_theIndividuals.size());
 }
 
 IIndividual* CPopulation::pick()
